Move util functions out of psn module
