+++ src/psn/mod.rs
@@ -1,16 +1,12 @@
+mod utils;
 mod parser;
 mod manifest_parser;
 
-use core::str;
-use std::fmt;
 use std::path::PathBuf;
 
-use hmac::{Hmac, Mac};
-use sha2::Sha256;
 use tokio::io::AsyncWriteExt;
 use tokio::sync::mpsc::Sender;
-
-type HmacSha256 = Hmac<Sha256>;
+use utils::{get_update_info_url, get_platform_variant, PlaformVariant};
 
 #[derive(Debug)]
 pub enum DownloadStatus {
@@ -66,7 +62,7 @@ impl UpdateInfo {
             Some(variant) => variant,
             None => return Err(UpdateError::InvalidSerial)
         };
-        let url = match choose_url(&title_id, platform_variant) {
+        let url = match get_update_info_url(&title_id, platform_variant) {
             Ok(url) => url,
             Err(err) => return Err(err)
         };
@@ -146,54 +142,6 @@ pub fn parse_title_id(title_id: &String) -> String {
         .trim()
         .replace("-", "") // strip the dash that some sites put in a title id, eg. BCES-xxxxx
         .to_uppercase();
-}
-
-#[derive(Clone, Copy, PartialEq, Debug)]
-pub enum PlaformVariant {
-    PS3,
-    PS4
-}
-
-impl fmt::Display for PlaformVariant {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{:?}", self)
-    }
-}
-
-fn get_platform_variant(title_id: &str) -> Option<PlaformVariant> {
-    if ["NP", "BL", "BC"].iter().any(|&prefix| { title_id.starts_with(prefix) }) {
-        return Some(PlaformVariant::PS3);
-    }
-
-    if title_id.starts_with("CUSA") {
-        return Some(PlaformVariant::PS4);
-    }
-
-    return None
-}
-
-fn choose_url(title_id: &str, platform_variant: PlaformVariant) -> Result<String, UpdateError> {
-    match platform_variant {
-        PlaformVariant::PS3 => {
-            Ok(format!("https://a0.ww.np.dl.playstation.net/tpl/np/{0}/{0}-ver.xml", title_id))
-        },
-        PlaformVariant::PS4 => {
-            let key = match hex::decode("AD62E37F905E06BC19593142281C112CEC0E7EC3E97EFDCAEFCDBAAFA6378D84") {
-                Ok(key) => key,
-                Err(_) => return Err(UpdateError::InvalidSerial),
-            };
-            let msg = format!("np_{0}", title_id);
-            let mut hasher = match HmacSha256::new_from_slice(&key) {
-                Ok(hasher) => hasher,
-                Err(_) => return Err(UpdateError::InvalidSerial)
-            };
-
-            hasher.update(msg.as_ref());
-            let hash_bytes = hasher.finalize().into_bytes();
-
-            Ok(format!("https://gs-sec.ww.np.dl.playstation.net/plo/np/{0}/{1:x}/{0}-ver.xml", title_id, hash_bytes))
-        }
-    }
 }
 
 #[derive(Clone)]
